app/execution/adaptive_task_executor.py
import logging
from app.agents.factory import AgentFactory
from app.knowledge.knowledge_graph import KnowledgeGraph
from app.entropy_management.entropy_manager import EntropyManager
from app.chat_with_ollama import ChatGPT
from app.workflow.workflow_engine import WorkflowEngine
from app.reinforcement_learning.advanced_rl import AdvancedRL
from app.human_feedback.feedback_interface import HumanFeedbackInterface
from typing import List, Dict, Any
import numpy as np
import asyncio
from tenacity import retry, stop_after_attempt, wait_exponential

logger = logging.getLogger(__name__)

class AdaptiveTaskExecutor:
    def __init__(self, agent_factory: AgentFactory, workflow_engine: WorkflowEngine, knowledge_graph: KnowledgeGraph, entropy_manager: EntropyManager, llm: ChatGPT, advanced_rl: AdvancedRL, human_feedback_interface: HumanFeedbackInterface):
        self.agent_factory = agent_factory
        self.workflow_engine = workflow_engine
        self.knowledge_graph = knowledge_graph
        self.entropy_manager = entropy_manager
        self.llm = llm
        self.advanced_rl = advanced_rl
        self.human_feedback_interface = human_feedback_interface
        self.state_dim = 100  # Set this to match the input_dim of AdvancedRL

    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
    async def execute_task(self, task: str):
        plan = await self._generate_plan(task)
        result = {"result": []}
        state = self._encode_state(task, "")
        for step in plan:
            action_probs = self.advanced_rl.get_action_probabilities(state)
            action = self._select_action(action_probs)
            step_result = await self._execute_action(action, step)
            result["result"].append(step_result)
            state = self._encode_state(task, step_result)
            await self._update_rl_model(state, action, step_result)
        return result

    async def _generate_plan(self, task: str):
        prompt = f"Generate a detailed, step-by-step plan for the following task: {task}"
        response = await self.llm.chat_with_ollama("You are a task planning expert.", prompt)
        return await self.workflow_engine.create_workflow({"task": task, "analysis": response})

    def _encode_state(self, task: str, step_result: str):
        task_embedding = self._get_embedding(task)
        result_embedding = self._get_embedding(step_result)
        return np.concatenate([task_embedding, result_embedding])

    def _get_embedding(self, text: str):
        # TODO: Implement a more sophisticated embedding method
        return np.random.rand(50)  # Return a 50-dimensional embedding

    def _select_action(self, action_probs):
        if self.entropy_manager.should_explore(action_probs):
            return np.random.choice(len(action_probs), p=action_probs)
        else:
            return np.argmax(action_probs)

    async def _execute_action(self, action: int, step: Dict[str, Any]):
        agent = await self.agent_factory.create_agent({"content": step["description"]})
        result = await agent.process_task({"content": step["description"], "action": action})
        if self._should_request_human_feedback(result):
            human_feedback = await self.human_feedback_interface.get_feedback(result)
            result = self._incorporate_human_feedback(result, human_feedback)
        return result

    def _should_request_human_feedback(self, result: Dict[str, Any]) -> bool:
        return result.get('confidence', 1.0) < 0.7

    def _incorporate_human_feedback(self, result: Dict[str, Any], feedback: str) -> Dict[str, Any]:
        result['human_feedback'] = feedback
        return result

    async def _update_rl_model(self, state: np.ndarray, action: int, result: Dict[str, Any]):
        reward = self._calculate_reward(result)
        self.advanced_rl.update(state, action, reward)

    def _calculate_reward(self, result: Dict[str, Any]) -> float:
        base_reward = result.get('success', False) * 1.0
        confidence_bonus = result.get('confidence', 1.0) * 0.5
        human_feedback_bonus = 0.5 if 'human_feedback' in result else 0
        return base_reward + confidence_bonus + human_feedback_bonus

    def _parse_plan(self, plan_text: str):
        return [step.strip() for step in plan_text.split('\n') if step.strip()]

    # Add other necessary methods